chore: split into files and separate logic
