+++ libs/rxidb/src/lib/database/index.ts
@@ -1,0 +1,2 @@
+export * from './connect-to-indexeddb';
+export * from './delete-indexeddb';
