+++ libs/rxidb/src/lib/object-store/index.ts
@@ -1,0 +1,3 @@
+export * from './get-object-store';
+export * from './crud';
+export * from './metadata';
