--- libs/rxidb/src/lib/rxidb.spec.ts
@@ -1,7 +1,0 @@
-import { rxidb } from './rxidb';
-
-describe('rxidb', () => {
-  it('should work', () => {
-    expect(rxidb()).toEqual('rxidb');
-  });
-});
